Added Console Commands
Added helper console commands to request opening the browser on the Operating System, going to the full LogId.
Added Client and Server commands, functions and ability to generate the LogId.
Added new TWeakPtr<UCapsaActorComponent> to the Subsystem. When running and set, this will provide the Server LogId.
Removed old variables for Looping looking for the class to auto-add the UCapsaActorComponent.